fix errors
fix implicit cast

don't import MatMulBnb4Quantizer at top level

sign type

fix doc

typename T -> class T

+++ onnxruntime/contrib_ops/cpu/quantization/blockwise_quant_block_bnb4.h
@@ -42,7 +42,7 @@ FORCEINLINE uint8_t QuantizeOneFP4(float x) {
   // that is difficult to noice if you add an extra
   // zero somewhere!
 
-  int sign = x < 0 ? 0b1000 : 0b0000;
+  uint8_t sign = x < 0 ? 0b1000 : 0b0000;
   x = fabsf(x);
   if (x > 0.29166667f) {
     if (x > 0.583333f) {
